Extact click handler into utility method
