+++ addon/components/frost-table.js
@@ -8,7 +8,7 @@ import {Component} from 'ember-frost-core'
 import {PropTypes} from 'ember-prop-types'
 
 import layout from '../templates/components/frost-table'
-import select from '../utils/selection'
+import {select} from '../utils/selection'
 import {ColumnPropType, ItemPropType, ItemsPropType} from 'ember-frost-table/typedefs'
 
 export default Component.extend({
